requirements.txt
# Required Python packages for Django backend
django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9.9
django-cors-headers>=4.4.0
openai>=1.30.0
python-dotenv>=1.0.1
gunicorn>=22.0.0
